+++ Code/VFSPrototype/VFSBase/Implementation/BlockList.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using VFSBase.Interfaces;
-using VFSBase.Persistance;
-using VFSBase.Persistance.Blocks;
+using VFSBase.Persistence;
+using VFSBase.Persistence.Blocks;
 
 namespace VFSBase.Implementation
 {
