+++ Code/VFSPrototype/VFSBase/Implementation/Folder.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using VFSBase.Interfaces;
-using VFSBase.Persistance;
-using VFSBase.Persistance.Blocks;
+
 
 namespace VFSBase.Implementation
 {
