+++ Code/VFSPrototype/VFSBase/Persistence/PositionCalculator.cs
@@ -1,4 +1,4 @@
-﻿namespace VFSBase.Persistance
+﻿namespace VFSBase.Persistence
 {
     internal class PositionCalculator
     {
