+++ Code/VFSPrototype/VFSBase/VFSBase.csproj
@@ -61,7 +61,7 @@
     <Compile Include="Interfaces\IFileSystemOptions.cs" />
     <Compile Include="Implementation\PathParser.cs" />
     <Compile Include="Interfaces\IFileSystem.cs" />
-    <Compile Include="Persistance\BinaryMathUtil.cs" />
+    <Compile Include="Persistence\BinaryMathUtil.cs" />
     <Compile Include="Implementation\FileSystemOptions.cs" />
     <Compile Include="Exceptions\NotFoundException.cs" />
     <Compile Include="Interfaces\IFileSystemTextManipulator.cs" />
@@ -69,11 +69,11 @@
     <Compile Include="Implementation\VFSFile.cs" />
     <Compile Include="Implementation\FileSystemTextManipulator.cs" />
     <Compile Include="Implementation\Folder.cs" />
-    <Compile Include="Persistance\BlockParser.cs" />
-    <Compile Include="Persistance\Blocks\EmptyBlock.cs" />
-    <Compile Include="Persistance\Blocks\IndirectNode.cs" />
-    <Compile Include="Persistance\PositionCalculator.cs" />
-    <Compile Include="Persistance\PositionCalculatorSettings.cs" />
+    <Compile Include="Persistence\BlockParser.cs" />
+    <Compile Include="Persistence\Blocks\EmptyBlock.cs" />
+    <Compile Include="Persistence\Blocks\IndirectNode.cs" />
+    <Compile Include="Persistence\PositionCalculator.cs" />
+    <Compile Include="Persistence\PositionCalculatorSettings.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Exceptions\VFSException.cs" />
   </ItemGroup>
