+++ Code/VFSPrototype/VFSBaseTests/BlockParserTest.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using VFSBase.Implementation;
-using VFSBase.Persistance;
-using VFSBase.Persistance.Blocks;
+using VFSBase.Persistence;
+using VFSBase.Persistence.Blocks;
 
 namespace VFSBaseTests
 {
